+++ CHANGELOG.md
@@ -23,8 +23,6 @@ Changes, both released, upcoming and in progress are kept here.
 
 - Rethink song link / URL / info panel.
 
-- Make album header horizontal
-
 ### Security
 - Get HTTPS Cert
 
@@ -33,6 +31,7 @@ Changes, both released, upcoming and in progress are kept here.
 ### Changed
 - Data description format allows linking to other albums or songs
 - Add dates to album card
+- Make album header horizontal on album page
 
 ### Fixed
 - Fix user IP in listens table...